nowe: folders

+++ _structures/Graph.cpp
@@ -115,9 +115,9 @@ Graph::Graph(string filePath) {
 
 void Graph::display() {
 
-    for (int i = 0; i < matrix.size(); ++i) {
+    for (size_t i = 0; i < matrix.size(); ++i) {
 
-        for (int j = 0; j < matrix.size(); ++j) {
+        for (size_t j = 0; j < matrix.size(); ++j) {
 
             cout << setw(5) << matrix[i][j] << " ";
         }
